fix(frontend): improve multi-enum setting dialog
The previous implementation of the multi-enum setting used a simple
OverlapGroup which wasn't user-friendly.

This commit refactors it to use a proper movable and centered dialog.
A new  widget is introduced to serve as the base for
this and potentially other future dialogs. This provides a better
user experience when editing settings with multiple options.

closes #30